Refactored file structure for SidebarWidget.
