+++ app/javascript/packs/SidebarWidget.jsx
@@ -1,5 +1,5 @@
 import { h, render } from 'preact';
-import SidebarWidget from '../sidebar-widget/SidebarWidget';
+import { SidebarWidget } from '../src/views/SidebarWidget';
 
 HTMLDocument.prototype.ready = new Promise(resolve => {
   if (document.readyState !== 'loading') {
