+++ app/javascript/src/views/SidebarWidget/SidebarWidget.jsx
@@ -1,8 +1,8 @@
 import { h, Component } from 'preact';
-import sendFollowUser from '../src/utils/sendFollowUser';
+import sendFollowUser from '../../utils/sendFollowUser';
 import SidebarUser from './SidebarUser';
 
-class SidebarWidget extends Component {
+export class SidebarWidget extends Component {
   constructor(props) {
     super(props);
     this.getSuggestedUsers = this.getSuggestedUsers.bind(this);
@@ -88,5 +88,3 @@ class SidebarWidget extends Component {
     return <div />;
   }
 }
-
-export default SidebarWidget;
