+++ app/javascript/src/views/SidebarWidget/index.js
@@ -1,0 +1,1 @@
+export * from './SidebarWidget';
